Move LED driver code to RAM and remove the explicit delay
The LED driver code must be in RAM, not in flash with XIP. When we write to flash we need to take the QSPI out of XIP mode, which breaks the 2nd core if it's running code with XIP.

The explicit loop delay was also running from flash, and I couldn't move it because it was a library function. Turns out we didn't really need it.